Initial commit of routes into separate classes.
This commit moves all of the API route registration into separate classes, one class per route. This will faciliate
the ease of adding endpoints to routes should additional CRUD operations be needed. It also ensures that routes are
following standard WP-API standards, such as extending the WP_REST_Controller (via an abstract class in some cases),
and by using the WP_REST_Server class to establish the method.

In addition, this commit includes three helper classes - Validator, Media_Synchronizer, and Post_Synchronizer, that
are individually used for performing the validation and synchronization actions that occur within an endpoint. We
should be able to split these into separate classes later on if their usage becomes more specialized, without affecting
the individual routes themselves.
